Updated and Reorganized folders and included Unittests
