+++ src/NPVCalculation.py
@@ -1,6 +1,4 @@
-import abc
-import math
-from CalculationInterface import CaculationInterface
+from Calculation_Core.CalculationInterface import CaculationInterface
 
 # Class Responsible for NPV Calculation
 # Takes the parameters of InitialInvestment,
